Django-Rest-Framework-Tutorial/requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
